Add Android signing, ProGuard & agent updates
Add Android release signing support (key.properties loading, fail-fast on missing file) and ProGuard rules for SQLCipher, secure storage, okhttp and Flutter. Expand .gitignore to exclude keystores, env files, build artifacts and platform ephemeral files; remove generated .metadata. Update README with release signing instructions, privacy and import details, API key storage note and other clarifications. Apply Android/iOS platform tweaks: INTERNET and POST_NOTIFICATIONS permissions, app name casing, orientation limits, and normalize generated Xcode/Flutter env paths. Improve agent code (ActionAgent, AnalysisAgent): formatting/refactors, Gemini availability checks, rate-limit/cooldown logic, reduced AI calls, better logging and error handling, portfolio insight cooldown, and minor robustness fixes. Also add utility/model scaffolding and small service adjustments.

+++ lib/core/models/monthly_archive.dart
@@ -37,15 +37,28 @@ class MonthlyArchive {
   /// Türkçe ay adı (Ocak, Şubat, ...)
   String get monthNameTr {
     const names = [
-      '', 'Ocak', 'Şubat', 'Mart', 'Nisan', 'Mayıs', 'Haziran',
-      'Temmuz', 'Ağustos', 'Eylül', 'Ekim', 'Kasım', 'Aralık',
+      '',
+      'Ocak',
+      'Şubat',
+      'Mart',
+      'Nisan',
+      'Mayıs',
+      'Haziran',
+      'Temmuz',
+      'Ağustos',
+      'Eylül',
+      'Ekim',
+      'Kasım',
+      'Aralık',
     ];
     return names[month];
   }
 
   /// Tasarruf oranı (0.0-1.0), gelir sıfırsa 0 döner
   double get savingsRate =>
-      totalIncome > 0 ? (totalSavings / totalIncome).clamp(0.0, 1.0) : 0.0;
+      totalIncome > 0
+          ? (totalSavings / totalIncome).clamp(0.0, 1.0).toDouble()
+          : 0.0;
 
   Map<String, dynamic> toDbMap() {
     return {
@@ -65,7 +78,8 @@ class MonthlyArchive {
 
   factory MonthlyArchive.fromDbMap(Map<String, dynamic> map) {
     final categoryJson =
-        jsonDecode(map['category_json'] as String? ?? '{}') as Map<String, dynamic>;
+        jsonDecode(map['category_json'] as String? ?? '{}')
+            as Map<String, dynamic>;
 
     return MonthlyArchive(
       id: map['id'] as String,
